Moved ReadDir* tests in their own file
